Rename ts_doc to tsdoc
