+++ tsdoc.ts
